scoutradioz-helpers v1.1.0 - typescript
