+++ scoutradioz-helpers/README.md
@@ -1,0 +1,6 @@
+## scoutradioz-helpers
+
+To update:
+1. `tsc`
+1. Edit `package.json` with a new version number
+1. `npm publish`
