+++ scoutradioz-helpers/types/index.d.ts
@@ -1,0 +1,10 @@
+import { Utilities } from '@firstteam102/scoutradioz-utilities';
+export declare class Helpers {
+    upload: any;
+    matchData: any;
+    /**
+     * Required:
+     * @param {object} utilities Already-configured @firstteam102/scoutradioz-utilities module
+     */
+    config: (utilities: Utilities) => void;
+}
